requirements.txt
Flask==2.0.2
pandas==1.3.3
SQLAlchemy==1.4.25
python-dotenv==0.19.1
Flask-SQLAlchemy==2.5.2
Authlib==0.15.4
sentry-sdk==1.5.1